requirements.txt
tensorflow==2.11.0
Pillow==9.3.0
matplotlib==3.6.2
